Update signal pages
