+++ src/data/signal-contentmap.js
@@ -6,11 +6,4 @@ export default {
     bannerType: 'fluidBlue',
     openGraph: 'signal-governance',
   },
-  "delegation-survey-2025": {
-    title: 'Delegation survey 2025',
-    formId: 6,
-    description: 'Help us understand delegation preferences and improve the staking experience.',
-    bannerType: 'fluidBlue',
-    openGraph: 'delegation-survey-2025',
-  },
 };
